修正：コントローラーの機能ごとの細分化、分離
